+++ info/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Last startup time using tick count.
 - Redundant startup application info check
+- Redundant hotfix info check
 
 ## 2024-06-01